+++ crates/wdl-analysis/src/validation.rs
@@ -15,6 +15,7 @@ use wdl_ast::TreeNode;
 use wdl_ast::VersionStatement;
 use wdl_ast::Whitespace;
 use wdl_ast::v1;
+use wdl_ast::v1::RequirementsItem;
 use wdl_grammar::Severity;
 use wdl_grammar::SyntaxKind;
 
@@ -539,6 +540,17 @@ impl Visitor for Validator {
     ) {
         for visitor in self.visitors.iter_mut() {
             visitor.requirements_section(diagnostics, reason, section);
+        }
+    }
+
+    fn requirements_item(
+        &mut self,
+        diagnostics: &mut Diagnostics,
+        reason: VisitReason,
+        item: &RequirementsItem,
+    ) {
+        for visitor in self.visitors.iter_mut() {
+            visitor.requirements_item(diagnostics, reason, item);
         }
     }