+++ crates/wdl-analysis/src/visitor.rs
@@ -49,6 +49,7 @@ use wdl_ast::v1::MetadataSection;
 use wdl_ast::v1::OutputSection;
 use wdl_ast::v1::ParameterMetadataSection;
 use wdl_ast::v1::Placeholder;
+use wdl_ast::v1::RequirementsItem;
 use wdl_ast::v1::RequirementsSection;
 use wdl_ast::v1::RuntimeItem;
 use wdl_ast::v1::RuntimeSection;
@@ -215,6 +216,15 @@ pub trait Visitor {
         diagnostics: &mut Diagnostics,
         reason: VisitReason,
         section: &RequirementsSection,
+    ) {
+    }
+
+    /// Visits a requirements item node.
+    fn requirements_item(
+        &mut self,
+        diagnostics: &mut Diagnostics,
+        reason: VisitReason,
+        item: &RequirementsItem,
     ) {
     }
 
@@ -466,9 +476,11 @@ pub(crate) fn visit<V: Visitor>(
             SyntaxKind::TaskHintsItemNode | SyntaxKind::WorkflowHintsItemNode => {
                 // Skip this node as it's part of a hints section
             }
-            SyntaxKind::RequirementsItemNode => {
-                // Skip this node as it's part of a requirements section
-            }
+            SyntaxKind::RequirementsItemNode => visitor.requirements_item(
+                diagnostics,
+                reason,
+                &RequirementsItem::cast(element.into_node().unwrap()).expect("should cast"),
+            ),
             SyntaxKind::RuntimeSectionNode => visitor.runtime_section(
                 diagnostics,
                 reason,